Move initpop to optimizer
