+++ biomass/estimation/__init__.py
@@ -1,3 +1,2 @@
-from .optimizer import Optimizer
-from .initpop import InitialPopulation
+from .optimizer import InitialPopulation, Optimizer
 from .search_util import convert_scale, initialize_search_param
