--- docs/api/estimation/index.rst
@@ -1,8 +1,0 @@
-Parameter estimation (:py:mod:`biomass.estimation`)
-===================================================
-
-.. toctree::
-   :maxdepth: 4
-
-   optimizer
-
