--- docs/api/estimation/optimizer.rst
@@ -1,5 +1,0 @@
-Optimizer (:py:mod:`biomass.estimation.optimizer`)
-==================================================
-
-.. autoclass:: biomass.estimation.Optimizer
-   :members:
