+++ docs/api/optimizer.rst
@@ -1,0 +1,8 @@
+Parameter estimation (:py:mod:`biomass.estimation.optimizer`)
+=============================================================
+
+.. autoclass:: biomass.estimation.Optimizer
+   :members: 
+
+.. autoclass:: biomass.estimation.InitialPopulation
+   :members:
